Reverted changes in pom.xml, combined EmptyArray and NullArray tests into NullAndEmptyArrayTest.java, and updated article as per the feedback
