--- core-java-modules/core-java-numbers-8/src/test/java/com/baeldung/nullandemptyarray/NullArrayUnitTest.java
@@ -1,16 +1,0 @@
-package com.baeldung.nullandemptyarray;
-
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
-import org.junit.jupiter.api.Test;
-
-public class NullArrayUnitTest {
-
-    @Test
-    public void givenNullArray_whenAccessLength_thenThrowsNullPointerException() {
-        int[] nullArray = null;
-        assertThrows(NullPointerException.class, () -> {
-            int length = nullArray.length;
-        });
-    }
-}
